Update manipulation demo
