+++ examples/manipulation-demo.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Robotec.AI
+# Copyright (C) 2025 Julia Jia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -9,26 +9,29 @@
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language goveself.rning permissions and
+# See the License for the specific language governing permissions and
 # limitations under the License.
 
 
 import logging
 from typing import List
 
+import rclpy
 from langchain_core.messages import BaseMessage, HumanMessage
 from langchain_core.tools import BaseTool
 from rai import get_llm_model
-from rai.agents.langchain.core import create_react_runnable
-from rai.communication.ros2 import wait_for_ros2_services, wait_for_ros2_topics
+from rai.agents.langchain.core import create_conversational_agent
 from rai.communication.ros2.connectors import ROS2Connector
 from rai.tools.ros2.manipulation import (
-    GetObjectPositionsTool,
     MoveObjectFromToTool,
     ResetArmTool,
 )
 from rai.tools.ros2.simple import GetROS2ImageConfiguredTool
-from rai_perception.tools import GetGrabbingPointTool
+from rai_perception import (
+    GetObjectGrippingPointsTool,
+    wait_for_perception_dependencies,
+)
+from rai_perception.tools.gripping_points_tools import GRIPPING_POINTS_TOOL_PARAM_PREFIX
 
 from rai_whoami.models import EmbodimentInfo
 
@@ -36,37 +39,60 @@ logger = logging.getLogger(__name__)
 
 
 def create_agent():
+    """Create and configure the manipulation agent.
+
+    GetObjectGrippingPointsTool auto-declares ROS2 parameters with defaults.
+    To override parameters for your deployment, declare them before tool initialization:
+
+        node.declare_parameter(f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.camera_topic", "/your/topic")
+        node.declare_parameter(f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.target_frame", "your_frame")
+        # ... etc
+
+    Or set them in launch files/YAML configs. See tool logs for current parameter values.
+    """
+    rclpy.init()
     connector = ROS2Connector(executor_type="single_threaded")
-
-    required_services = ["/segmentation", "/detection"]
-    required_topics = ["/color_image5", "/depth_image5", "/color_camera_info5"]
-    wait_for_ros2_services(connector, required_services)
-    wait_for_ros2_topics(connector, required_topics)
-
     node = connector.node
-    node.declare_parameter("conversion_ratio", 1.0)
+
+    # Set ROS2 parameters to match robot/simulation topic and frame names
+    # For O3DE simulation, these are the default values
+    # To use different values, change these or set them in a launch file
+    node.declare_parameter(
+        f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.camera_topic", "/color_image5"
+    )
+    node.declare_parameter(
+        f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.depth_topic", "/depth_image5"
+    )
+    node.declare_parameter(
+        f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.camera_info_topic", "/color_camera_info5"
+    )
+    node.declare_parameter(
+        f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.target_frame", "panda_link0"
+    )
+    node.declare_parameter(
+        f"{GRIPPING_POINTS_TOOL_PARAM_PREFIX}.source_frame", "RGBDCamera5"
+    )
+
+    # Initialize tools - GetObjectGrippingPointsTool will use the parameters we just set
+    gripping_points_tool = GetObjectGrippingPointsTool(connector=connector)
+    config = gripping_points_tool.get_config()
 
     tools: List[BaseTool] = [
-        GetObjectPositionsTool(
-            connector=connector,
-            target_frame="panda_link0",
-            source_frame="RGBDCamera5",
-            camera_topic="/color_image5",
-            depth_topic="/depth_image5",
-            camera_info_topic="/color_camera_info5",
-            get_grabbing_point_tool=GetGrabbingPointTool(connector=connector),
+        gripping_points_tool,
+        MoveObjectFromToTool(
+            connector=connector, manipulator_frame=config["target_frame"]
         ),
-        MoveObjectFromToTool(connector=connector, manipulator_frame="panda_link0"),
-        ResetArmTool(connector=connector, manipulator_frame="panda_link0"),
-        GetROS2ImageConfiguredTool(connector=connector, topic="/color_image5"),
+        ResetArmTool(connector=connector, manipulator_frame=config["target_frame"]),
+        GetROS2ImageConfiguredTool(connector=connector, topic=config["camera_topic"]),
     ]
+
+    wait_for_perception_dependencies(connector, tools)
 
     llm = get_llm_model(model_type="complex_model", streaming=True)
     embodiment_info = EmbodimentInfo.from_file(
         "examples/embodiments/manipulation_embodiment.json"
     )
-
-    agent = create_react_runnable(
+    agent = create_conversational_agent(
         llm=llm,
         tools=tools,
         system_prompt=embodiment_info.to_langchain(),
